requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
mysqlclient>=2.2.0
python-decouple>=3.8
Pillow>=10.0.0
reportlab>=4.0.0
razorpay>=1.4.0
django-filter>=23.5
pyjwt>=2.8.0
python-dateutil>=2.8.2
